requirement.txt
fastapi[standard]>=0.113.0,<0.114.0
pydantic>=2.7.0,<3.0.0
uvicorn[standard]==0.23.0
psycopg2-binary==2.9.9
